backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wikis-backend"
version = "0.1.0"
description = "Wikis — AI-powered repository documentation generator"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    # LLM orchestration
    "langchain_core>=0.3.49",
    "langchain>=0.3.0",
    "langchain_community>=0.3.7",
    "langgraph>=1.0.0",
    "deepagents>=0.3.0",
    # LLM providers
    "langchain-openai>=0.3.0",
    "langchain-anthropic>=1.0.0",
    "openai>=1.55.0",
    "anthropic>=0.57.0",
    # Search & indexing
    "faiss-cpu>=1.7.0",
    "sentence-transformers>=2.2.0",
    "rank_bm25",
    # Code parsing
    "tree-sitter-language-pack==0.9.1",
    # Git
    "pygithub==2.3.0",
    # Database
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29.0",
    "aiosqlite>=0.20.0",
    # Auth
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.27.0",
    # MCP
    "mcp>=1.0.0",
    # SSE (also a transitive dep of mcp, declared explicitly for routes.py)
    "sse-starlette>=2.0.0",
    # Utilities
    "requests>=2.25.0",
    "tiktoken>=0.5.0",
    "langchain_text_splitters>=0.3.0",
]

[project.optional-dependencies]
ollama = ["langchain-ollama>=0.3.0"]
gemini = ["langchain-google-genai>=2.0.0"]
bedrock = ["langchain-aws>=0.2.0"]
all-providers = ["wikis-backend[ollama,gemini,bedrock]"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "pytest-cov>=5.0",
    "ruff>=0.4.0",
]

[project.scripts]
wikis-mcp = "mcp_server.server:main"

[tool.setuptools.packages.find]
include = ["app*", "mcp_server*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "UP",     # pyupgrade
    "S",      # flake8-bandit (security)
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # function call in default arg (FastAPI Depends() pattern)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S", "B017"]  # disable security checks + allow broad exception asserts in tests
"app/core/*" = [
    "E402",   # import-not-at-top (conditional imports in core)
    "E741",   # ambiguous-variable-name (single-letter vars in algorithms)
    "F811",   # redefined-while-unused (conditional imports)
]

[tool.ruff.format]
quote-style = "double"
